Move `ui_open` into its own `util` module.
