+++ src/python/pants/backend/jvm/tasks/BUILD
@@ -185,13 +185,12 @@ python_library(
     ':classpath_products',
     ':ivy_task_mixin',
     ':nailgun_task',
-    'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/jvm:ivy_utils',
+    'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:exceptions',
-    'src/python/pants/binaries:binary_util',
     'src/python/pants/invalidation',
+    'src/python/pants/util:desktop',
     'src/python/pants/util:dirutil',
-    'src/python/pants/util:memo',
     'src/python/pants/util:strutil',
   ],
 )
@@ -302,11 +301,13 @@ python_library(
   name = 'coverage',
   sources = globs('coverage/*.py'),
   dependencies = [
-    '3rdparty/python/twitter/commons:twitter.common.dirutil',
-    ':classpath_util',
-    'src/python/pants/base:build_environment',
-    'src/python/pants/binaries:binary_util',
+    '3rdparty/python/twitter/commons:twitter.common.collections',
+    'src/python/pants/backend/jvm/targets:jvm',
+    'src/python/pants/base:exceptions',
+    'src/python/pants/util:contextutil',
+    'src/python/pants/util:desktop',
     'src/python/pants/util:dirutil',
+    'src/python/pants/util:meta',
     'src/python/pants/util:strutil',
   ],
 )
@@ -333,14 +334,14 @@ python_library(
   dependencies = [
     '3rdparty/python:six',
     ':classpath_util',
+    ':coverage',
     ':jvm_task',
     ':jvm_tool_task_mixin',
+    ':reports',
     'src/python/pants/backend/jvm/subsystems:jvm_platform',
     'src/python/pants/backend/jvm/subsystems:shader',
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/jvm/targets:jvm',
-    'src/python/pants/backend/jvm/tasks:coverage',
-    'src/python/pants/backend/jvm/tasks:reports',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:workunit',
     'src/python/pants/binaries:binary_util',
@@ -478,6 +479,7 @@ python_library(
     ':jvm_task',
     'src/python/pants/base:exceptions',
     'src/python/pants/binaries:binary_util',
+    'src/python/pants/util:desktop',
     'src/python/pants/util:dirutil',
   ],
 )
