+++ src/python/pants/backend/project_info/tasks/BUILD
@@ -120,16 +120,19 @@ python_library(
   name = 'ide_gen',
   sources = ['ide_gen.py'],
   dependencies = [
+    '3rdparty/python:pathspec',
     '3rdparty/python/twitter/commons:twitter.common.collections',
+    'src/python/pants/backend/jvm/subsystems:scala_platform',
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/jvm/targets:scala',
     'src/python/pants/backend/jvm/tasks:classpath_products',
     'src/python/pants/backend/jvm/tasks:ivy_task_mixin',
     'src/python/pants/backend/jvm/tasks:nailgun_task',
     'src/python/pants/base:build_environment',
-    'src/python/pants/build_graph',
+    'src/python/pants/base:build_file',
     'src/python/pants/base:exceptions',
-    'src/python/pants/binaries:binary_util',
+    'src/python/pants/build_graph',
+    'src/python/pants/util:desktop',
     'src/python/pants/util:dirutil',
   ],
 )
@@ -158,10 +161,13 @@ python_library(
   ],
   dependencies = [
     ':ide_gen',
-    'src/python/pants/backend/jvm/targets:java',
+    'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:build_environment',
+    'src/python/pants/base:exceptions',
     'src/python/pants/base:generator',
-    'src/python/pants/scm:git',
+    'src/python/pants/task',
+    'src/python/pants/util:contextutil',
+    'src/python/pants/util:desktop',
     'src/python/pants/util:dirutil',
   ],
 )
