+++ src/python/pants/binaries/binary_util.py
@@ -8,7 +8,6 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 import logging
 import os
 import posixpath
-import subprocess
 from contextlib import contextmanager
 
 from twitter.common.collections import OrderedSet
@@ -155,7 +154,7 @@ class BinaryUtil(object):
           'No urls are defined for the --pants-support-baseurls option.')
     downloaded_successfully = False
     accumulated_errors = []
-    for baseurl in OrderedSet(self._baseurls):  # Wrap in OrderedSet because duplicates are wasteful.
+    for baseurl in OrderedSet(self._baseurls):  # De-dup URLS: we only want to try each URL once.
       url = posixpath.join(baseurl, binary_path)
       logger.info('Attempting to fetch {name} binary from: {url} ...'.format(name=name, url=url))
       try:
@@ -258,37 +257,3 @@ def safe_args(args,
         yield create_argfile(fp)
   else:
     yield args
-
-
-def _mac_open(files):
-  subprocess.call(['open'] + list(files))
-
-
-def _linux_open(files):
-  cmd = "xdg-open"
-  if not _cmd_exists(cmd):
-    raise TaskError("The program '{}' isn't in your PATH. Please install and re-run this "
-                    "goal.".format(cmd))
-  for f in list(files):
-    subprocess.call([cmd, f])
-
-
-# From: http://stackoverflow.com/questions/377017/test-if-executable-exists-in-python
-def _cmd_exists(cmd):
-  return subprocess.call(["/usr/bin/which", cmd], shell=False, stdout=subprocess.PIPE,
-                         stderr=subprocess.PIPE) == 0
-
-_OPENER_BY_OS = {
-  'darwin': _mac_open,
-  'linux': _linux_open
-}
-
-
-def ui_open(*files):
-  """Attempts to open the given files using the preferred native viewer or editor."""
-  if files:
-    osname = os.uname()[0].lower()
-    if not osname in _OPENER_BY_OS:
-      print('Sorry, open currently not supported for ' + osname)
-    else:
-      _OPENER_BY_OS[osname](files)
